backend/requirements.txt
redis==5.0.8
aiohttp==3.10.5
python-socketio==5.11.4
